More fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@13354 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/sound/loader/wav/wavfile.cpp
@@ -113,7 +113,7 @@ public:
     SCF_CONSTRUCT_IBASE(p);
     SCF_CONSTRUCT_EMBEDDED_IBASE(scfiComponent);
   }
-  virtual csPtr<iSoundData> LoadSound(void *Buffer, unsigned long Size);
+  virtual csPtr<iSoundData> LoadSound(void *Buffer, uint32 Size);
 };
 
 SCF_IMPLEMENT_IBASE(csSoundLoader_WAV)
